Split up and rename prefill contact details hook
